[EORL-8] tempReservation CRUD 추가

+++ src/main/java/com/eorl/service/member/MemberServiceImpl.java
@@ -2,7 +2,7 @@ package com.eorl.service;
 
 import com.eorl.domain.member.member.Member;
 import com.eorl.repository.MemberRepository;
-import java.util.NoSuchElementException;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DuplicateKeyException;
@@ -12,8 +12,7 @@ import org.springframework.stereotype.Component;
 //@Service
 @Component
 @RequiredArgsConstructor
-public class MemberServiceImpl implements MemberService {
-
+public class MemberServiceImpl implements  MemberService {
     private final MemberRepository memberRepository;
 
     @Override
@@ -29,7 +28,7 @@ public class MemberServiceImpl implements MemberService {
     }
 
     @Override
-    public Member findByMemberId(int memberId) {
+    public Member findByMemberId(Long memberId) {
         Member member = memberRepository.findByMemberId(memberId);
         if (member == null) {
             throw new NoSuchElementException("아이디가 '" + memberId + "' 멤버는 존재하지 않습니다.");
@@ -52,7 +51,7 @@ public class MemberServiceImpl implements MemberService {
     }
 
     @Override
-    public int updateMemberAuthentication(String phoneNumber, int memberId) {
+    public int updateMemberAuthentication(String phoneNumber, Long memberId) {
         if (memberRepository.findByMemberId(memberId) == null) {
             throw new NoSuchElementException("아이디가 '" + memberId + "' 멤버는 존재하지 않습니다.");
         }
@@ -64,7 +63,7 @@ public class MemberServiceImpl implements MemberService {
     }
 
     @Override
-    public void deleteMember(int memberId) {
+    public void deleteMember(Long memberId) {
 
         if (memberRepository.findByMemberId(memberId) == null) {
             throw new NoSuchElementException("아이디가 '" + memberId + "' 멤버는 존재하지 않습니다.");
